+++ src/components/LayerCollection/LayerPOLine.js
@@ -95,6 +95,8 @@ const parseErrorMessage = (code) => {
 };
 
 const FIELD_ARRAYS_TO_HYDRATE = ['locations'];
+const MISSING_AFFILIATION_ERROR_CODE = 'missingAffiliation';
+const INVALID_TOKEN_MESSAGE = 'Invalid token';
 
 function LayerPOLine({
   history,
@@ -191,7 +193,11 @@ function LayerPOLine({
           let messageCode = get(ERROR_CODES, response.errors[0].code);
 
           if (!messageCode) {
-            messageCode = parseErrorMessage(response.errors[0].message);
+            const message = response.errors[0].message;
+
+            messageCode = message?.includes(INVALID_TOKEN_MESSAGE)
+              ? MISSING_AFFILIATION_ERROR_CODE
+              : parseErrorMessage(message);
           }
 
           sendCallout({